Continued work on general connection logic and virtual functions for connection events
