+++ Pixel-Game-Core/CMakeLists.txt
@@ -4,14 +4,14 @@ target_sources(Pixel-Game-Core PRIVATE
 	src/Chunk.cpp
 	src/Chunk.h
 	src/Element.h
-	src/PixelGameMessage.h
+	
 	src/PixelRigidBody.cpp
 	src/PixelRigidBody.h
 	src/VectorHash.h
 	src/World.cpp
 	src/World.h
 
-)
+ "src/PixelNetworking.h")
 
 target_include_directories(Pixel-Game-Core PUBLIC
 	src
